+++ packages/sizuku/src/generator/valibot/generator/infer-input.ts
@@ -1,4 +1,4 @@
-import { capitalize } from '../../../shared/utils/capitalize.js'
+import { capitalize } from "../../../shared/utils/index.js";
 
 /**
  * @param name
