+++ packages/sizuku/src/generator/valibot/generator/valibot-code.ts
@@ -1,4 +1,3 @@
-import type { Schema } from '../../../shared/types.js'
 import { inferInput } from './infer-input.js'
 import { valibot } from './valibot.js'
 
@@ -8,7 +7,14 @@ import { valibot } from './valibot.js'
  * @param type
  * @returns
  */
-export function valibotCode(schema: Schema, comment: boolean, type: boolean): string {
+export function valibotCode(schema: {
+  name: string
+  fields: {
+    name: string
+    definition: string
+    description?: string
+  }[]
+}, comment: boolean, type: boolean): string {
   const valibotSchema = valibot(schema, comment)
 
   if (type) {
