+++ packages/sizuku/src/generator/zod/generator/zod-code.ts
@@ -1,4 +1,3 @@
-import type { Schema } from '../../../shared/types.js'
 import { infer } from './infer.js'
 import { zod } from './zod.js'
 
@@ -10,7 +9,14 @@ import { zod } from './zod.js'
  * @param config - The configuration for the code generation.
  * @returns The generated Zod code.
  */
-export function zodCode(schema: Schema, comment: boolean, type: boolean): string {
+export function zodCode(schema: {
+  name: string
+  fields: {
+    name: string
+    definition: string
+    description?: string
+  }[]
+}, comment: boolean, type: boolean): string {
   const zodSchema = zod(schema, comment)
   if (type) {
     const zInfer = infer(schema.name)
