--- packages/sizuku/src/shared/types.ts
@@ -1,10 +1,0 @@
-export type SchemaField = {
-  name: string
-  definition: string
-  description?: string
-}
-
-export type Schema = {
-  name: string
-  fields: SchemaField[]
-}
